Made edits and changes of avatar persistent

+++ public/js/tasklist.js
@@ -5,6 +5,7 @@ $(document).ready(function() {
 	initializePage();
 })
 
+
 /*
  * Function that is called when the document is ready.
  */
@@ -28,6 +29,7 @@ function initializePage() {
     $('#editTime').timepicker({
         minuteStep: 5
     });
+    $('#taskTime').val('');
 
 
     // Add any additional listeners here
@@ -44,6 +46,11 @@ function initializePage() {
 }
 
 
+/*
+ * Function to add new tasks to lists.  It takes the values found in fields and
+ * appends a new list item with appropriate information.  A new data json is
+ * created and stored into the file
+ */
 function addList(e) {
     e.preventDefault();
 
@@ -58,14 +65,35 @@ function addList(e) {
     var taskID = "task" + task;
 
     var deadline = $("#taskDeadlineCheck").prop('checked');
+    if (!deadline) {
+        time = '';
+    }
 
     var fields = ["• Task Name", "• Deadline Date", "• Deadline Time"];
     var warn = "You are missing the following field(s):";
 
-    var data = {"id": task, "name": name, "date": date, "time": time};
-    data.type="post";
-    $.post('/home', data, function(res) {});
+    var today = new Date();
+    var month = today.getMonth() + 1;
+    var mm = month < 10 ? "0" + month.toString() : month.toString();
+    var dd =  today.getDate().toString();
+    var yyyy = today.getFullYear().toString();
+    var createDate = mm + "/" + dd + "/" + yyyy;
 
+    var curHour = today.getHours();
+    var hh = curHour === 0 ? "12" : (curHour > 12 ? (curHour - 12).toString() : curHour.toString());
+    var min = today.getMinutes().toString();
+    var meridiem = curHour < 12 ? " AM" : " PM";
+    var createTime = hh + ":" + min + meridiem;
+
+    var data = {
+        "id": task,
+        "name": name,
+        "deadline": deadline,
+        "date": date,
+        "time": time,
+        "created": createDate + ", " + createTime
+    };
+    data.type= "post";
 
     if (deadline) {
         if (name.length > 0 && date.length > 0 && time.length > 0) {
@@ -74,12 +102,14 @@ function addList(e) {
                 '<h4 class="list-group-item-heading"><i class="glyphicon glyphicon-edit"></i> ' + name + '</h4>' +
                 '<p>Deadline: ' + date + ', ' + time + '</p></li>' +
                 '</a>');
-            $("#taskModal").modal('hide');
-            $("#taskName").val('');
-            $("#taskDate").val('');
-            $("#taskTime").val('');
-            $("#taskDeadlineCheck").prop("checked", false);
+            $('#taskModal').modal('hide');
+            $('#taskName').val('');
+            $('#taskDate').val('');
+            $('#taskTime').val('');
+            $('#taskDeadlineCheck').prop("checked", false);
             $('#taskDeadlineFields').hide();
+
+            $.post('/home', data, function(res) {});
         }
         else {
             if (name.length <= 0) {
@@ -110,28 +140,39 @@ function addList(e) {
             $("#taskTime").val('');
             $("#taskDeadlineCheck").prop("checked", false);
             $('#taskDeadlineFields').hide();
+
+            $.post('/home', data, function(res) {});
         }
     }
 }
 
 
+/*
+ * Function to clean up after CANCEL is hit for the ADD TASK MODAL
+ */
 function cancelAdd(e) {
-    $("#taskModal").modal('hide');
-    $("#taskName").val('');
-    $("#taskDate").val('');
-    $("#taskTime").val('');
-    $("#taskDeadlineCheck").prop("checked", false);
+    $('#taskModal').modal('hide');
+    $('#taskName').val('');
+    $('#taskDate').val('');
+    $('#taskTime').val('');
+    $('#taskDeadlineCheck').prop("checked", false);
     $('#taskDeadlineFields').hide();
 }
 
 
+/*
+ * Function to display appropriate information in the EDIT TASK MODAL;
+ * This is where the taskID is injected into a class where selecting options
+ * like DONE, DELETE, and SAVE are able to locate the appropriate list item
+ * to change.
+ */
 function openEdit(obj) {
     var task = $(obj);
 
     var taskID = task.find(".list-group-item");
     var taskHead = task.find(".list-group-item-heading");
     var taskName = taskHead.text().substr(1);
-    $("#editTask").val(taskName);
+    $('#editTask').val(taskName);
 
     var taskBody = task.find("p");
     var taskDead = taskBody.text().substr("Deadline: ".length);
@@ -139,17 +180,20 @@ function openEdit(obj) {
 
     // If it says "None, there is no deadline."  Otherwise, DEADLINE.
     if(taskDead === "None") {
-        $("#editDeadlineCheck").prop("checked", false);
+        $('#editDeadlineCheck').prop("checked", false);
         $('#editDeadlineFields').hide();
-        $("#editDate").val('');
-        $("#editTime").val('');
+        $('#editDate').val('');
+        $('#editDate').attr("placeholder", "Date");
+        $('#editTime').val('');
+        $('#editTime').attr("placeholder", "Time");
+
     }
     else {
-        $("#editDeadlineCheck").prop("checked", true);
+        $('#editDeadlineCheck').prop("checked", true);
         $('#editDeadlineFields').show();
 
-        $("#editDate").val(taskDateTime[0]);
-        $("#editTime").val(taskDateTime[1]);
+        $('#editDate').val(taskDateTime[0]);
+        $('#editTime').val(taskDateTime[1]);
     }
 
     var taskHolder = $('#editModal').find('#edit');
@@ -157,6 +201,13 @@ function openEdit(obj) {
 }
 
 
+/* If a person edits an entire task, though, the creation date would have to change, but
+ * I'm lazy, so no.
+ *
+ * Function to EDIT the task; using the injected taskID, the modal locates the task that
+ * launched it and proceeds to gather all the appropriate information from the fields to
+ * reflect the edits in the list view.
+ * */
 function editTask(e) {
     e.preventDefault();
 
@@ -168,7 +219,7 @@ function editTask(e) {
 
     taskTitle.html('<i class="glyphicon glyphicon-edit"></i> ' + newTitle);
 
-    var deadline = $("#editDeadlineCheck").prop('checked');
+    var deadline = $('#editDeadlineCheck').prop('checked');
 
     var taskBody = task.find("p");
     var newDate = $('#editDate').val();
@@ -180,37 +231,71 @@ function editTask(e) {
     }
     else {
         taskBody.text("Deadline: None");
+        newDate = "";
+        newTime = "";
     }
+
+    var data = {
+        "id": taskID.substr("task".length),
+        "name": newTitle,
+        "deadline": deadline,
+        "date": newDate,
+        "time": newTime
+    };
+    data.type = "edit";
+    $.post('/home', data, function (res) {});
 
     $('#editModal').find('#edit').removeClass(taskID);
     $('#editModal').modal('hide');
 }
 
+
+/*
+ * Function to DELETE the task; using the injected taskID, the modal locates the task that
+ * launched it to remove
+ *
+ * TODO: Find some way to get it to remove from the json because that isn't working
+ */
 function deleteTask(e) {
     e.preventDefault();
-    var obj = $(this);
 
     var taskID = $('#editModal').find('#edit').attr('class');
     var task = $('#' + taskID);
 
-    var data = {type: "delete"};
+    if (confirm("Are you sure you want to delete this task?")) {
 
-    // taskID
-    data.id = taskID.substr("task".length);
+      var data = {type: "delete"};
 
-    //add new
-    var taskHead = task.find(".list-group-item-heading");
-    var taskName = taskHead.text().substr(1);
-    console.log("taskName: " + taskName);
-    // end new
+      // taskID
+      data.id = taskID.substr("task".length);
 
-    console.log(taskID.substr("task".length));
-    data.taskName = taskHead.text().substr(1);
-    $.post('/home', data, function (res) {});
+      //add new
+      var taskHead = task.find(".list-group-item-heading");
+      var taskName = taskHead.text().substr(1);
+      console.log("taskName: " + taskName);
+      // end new
 
-    console.log(obj);
+      console.log(taskID.substr("task".length));
+      data.taskName = taskHead.text().substr(1);
+      $.post('/home', data, function (res) {});
+
+      $("#editModal").modal('hide');
+    }
+
+    task.hide();
+    $('#editModal').find('#edit').removeClass(taskID);
+    $('#editModal').modal('hide');
 }
 
+
+/*
+ * Function to FINISH the task; using the injected taskID, the modal locates the task that
+ * launched it to remove it from the list and to increment the user's points by the appropriate
+ * amount of points.
+ *
+ * TODO: Still have yet to increment points or figured out how to remove a task from json or to
+ *       edit the JSON information
+ */
 function doneTask(e) {
     e.preventDefault();
     var obj = $(this);
@@ -233,10 +318,19 @@ function doneTask(e) {
     data.taskName = taskHead.text().substr(1);
     $.post('/home', data, function (res) {});
 
-    console.log(obj);
+    task.hide();
+    $("#editModal").modal('hide');
+    $("#editName").val('');
+    $("#editDate").val('');
+    $("#editTime").val('');
+    $("#editDeadlineCheck").prop("checked", false);
+    $('#editDeadlineFields').hide();
 }
 
 
+/*
+ * Function to clean up the EDIT MODAL after CANCEL is hit
+ */
 function cancelEdit(e) {
     var taskID = $('#editModal').find('#edit').attr('class');
     $('#editModal').find('#edit').removeClass(taskID);
@@ -248,7 +342,10 @@ function cancelEdit(e) {
     $('#editDeadlineFields').hide();
 }
 
-/* */
+/*
+ * Function to show and hide the DATE and TIME fields in the modal when the
+ * DEADLINE? checkbox is toggled
+ */
 function toggleDeadline(classID, obj) {
 	var input = $(obj);
 	if( input.prop('checked')) {
